+++ generators/app/index.js
@@ -1,12 +1,8 @@
 'use strict';
 
 const cp = require('child_process');
-const path = require('path');
 
-const camelCase = require('camelcase');
-const chalk = require('chalk');
 const Generator = require('yeoman-generator');
-const yosay = require('yosay');
 
 let username = ' ';
 
@@ -19,124 +15,15 @@ try {
 
 module.exports = class extends Generator {
   prompting() {
-    // Have Yeoman greet the user.
-    this.log(
-      yosay(
-        `Behold the almighty ${chalk.red('generator-cheminfo')} generator!`,
-      ),
-    );
-
-    var prompts = [
-      {
-        type: 'input',
-        name: 'name',
-        message: 'Your project name',
-        default: path.basename(this.destinationRoot()), // Default to current folder name
-      },
-      {
-        type: 'input',
-        name: 'org',
-        message: 'GitHub organization',
-        default: 'cheminfo',
-      },
-      {
-        type: 'input',
-        name: 'userName',
-        message: 'Your name',
-        default: username.substring(0, username.length - 1),
-        store: true,
-      },
-      {
-        type: 'input',
-        name: 'description',
-        message: 'Your package description',
-      },
-    ];
-
-    return this.prompt(prompts).then(
-      function(props) {
-        // To access props later use this.props.name;
-        this.props = props;
-      }.bind(this),
-    );
+    this.log('The following generators are available:');
+    this.log('- yo cheminfo:module');
+    this.log('- yo cheminfo:typescript');
+    this.log('- yo cheminfo:lerna-module');
+    this.log('- yo cheminfo:react-frontend');
+    this.log('- yo cheminfo:commonjs');
   }
 
-  writing() {
-    const date = new Date();
-    const day = date.getDate();
-    const month = date.getMonth();
-    const year = date.getFullYear();
-    const camelName = camelCase(this.props.name);
-    const prefix = this.props.org === 'mljs' ? 'ml-' : '';
-    const includes = {
-      npmName: prefix + this.props.name,
-      name: this.props.name,
-      org: this.props.org,
-      userName: this.props.userName,
-      description: this.props.description,
-      date: year + '-' + month + '-' + day,
-      year: year,
-      camelName: camelName,
-    };
+  writing() {}
 
-    this.fs.copy(
-      this.templatePath('gitignore'),
-      this.destinationPath('.gitignore'),
-    );
-    this.fs.copy(
-      this.templatePath('eslintrc.yml'),
-      this.destinationPath('.eslintrc.yml'),
-    );
-
-    this.fs.copyTpl(
-      this.templatePath('index'),
-      this.destinationPath('src/index.js'),
-      includes,
-    );
-    this.fs.copyTpl(
-      this.templatePath('test'),
-      this.destinationPath('src/__tests__/test.js'),
-      includes,
-    );
-    this.fs.copy(
-      this.templatePath('npmignore'),
-      this.destinationPath('src/.npmignore'),
-    );
-    this.fs.copy(this.templatePath('npmrc'), this.destinationPath('.npmrc'));
-    this.fs.copyTpl(
-      this.templatePath('nodejs.yml'),
-      this.destinationPath('.github/workflows/nodejs.yml'),
-      includes,
-    );
-
-    this.fs.copyTpl(
-      this.templatePath('LICENSE'),
-      this.destinationPath('LICENSE'),
-      includes,
-    );
-    this.fs.copyTpl(
-      this.templatePath('package'),
-      this.destinationPath('package.json'),
-      includes,
-    );
-    this.fs.copyTpl(
-      this.templatePath('README.md'),
-      this.destinationPath('README.md'),
-      includes,
-    );
-  }
-
-  install() {
-    let deps = [
-      'eslint',
-      'eslint-config-cheminfo',
-      'eslint-plugin-import',
-      'eslint-plugin-jest',
-      'eslint-plugin-prettier',
-      'jest',
-      'prettier',
-    ];
-
-    this.npmInstall(deps, { 'save-dev': true });
-  }
+  install() {}
 };