+++ generators/commonjs/templates/index.js
@@ -1,0 +1,13 @@
+'use strict';
+
+/**
+ * Returns a very important number
+ * @return {number}
+ */
+function myCommonJS() {
+  return 42;
+}
+
+module.exports = {
+  myCommonJS,
+};